requirements.txt
numpy
scipy
times
python-csv
pandas
sgp4
skyfield
tensorflow==2.2.0
sklearn
pyrcca
seaborn
openpyxl
cython
eif
mpld3
plotly
pygame
pydmd